+++ malaria_web/admin.py
