+++ malaria_web/models.py
