+++ malaria_web/templates/malaria_web/delete_post.html
@@ -10,7 +10,7 @@
 <div class="well">
     <h2 class="text-center inverted-font">Delete Post</h2>
     <br>
-    <form action="{% url 'malaria:delete_post' post_id %}" method="post">
+    <form action="{% url 'malaria_web:delete_post' post_id %}" method="post">
         {% csrf_token %}
         <div class="form-group">
             <label class="col-sm-offset-2 inverted-font">Are you sure you want to delete this Post?</label>
@@ -18,7 +18,7 @@
         <div class="form-group">
             <div class="col-sm-offset-2">
                 <button type="submit" class="btn btn-danger">Yes, delete this Post</button>
-                <a href="{% url 'malaria:list_posts' %}" class="btn btn-default">Cancel</a>
+                <a href="{% url 'malaria_web:list_posts' %}" class="btn btn-default">Cancel</a>
             </div>
         </div>
     </form>
