+++ malaria_web/templates/malaria_web/view_post.html
@@ -29,9 +29,9 @@
     <br>
     <div class="row">
         <div class="col-sm-12 col-sm-offset-2">
-            <a href="{% url 'malaria:list_posts' %}" class="btn btn-default">Back to Post List</a>
-            <a href="{% url 'malaria:edit_post' post.id %}" class="btn btn-info">Edit</a>
-            <a href="{% url 'malaria:delete_post' post.id %}" class="btn btn-danger">Delete</a>
+            <a href="{% url 'malaria_web:list_posts' %}" class="btn btn-default">Back to Post List</a>
+            <a href="{% url 'malaria_web:edit_post' post.id %}" class="btn btn-info">Edit</a>
+            <a href="{% url 'malaria_web:delete_post' post.id %}" class="btn btn-danger">Delete</a>
         </div>
     </div>
 </div>
