+++ webhub/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from malaria.models import Post, RevPost
+from malaria_web.models import Post, RevPost
 from peacetrack.models import *
 from webhub.models import *
